+++ NewZhongYan/NewZhongYan/SKAppConfiguration.m
@@ -117,9 +117,9 @@
 }
 
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    return 88;
+    return 88.;
 }
 
 #pragma mark -tableViewDelegate
